Fix build use __u64 instead of __uint128_t

+++ arch/arm64/include/uapi/asm/sigcontext.h
@@ -50,7 +50,7 @@ struct fpsimd_context {
 	struct _aarch64_ctx head;
 	__u32 fpsr;
 	__u32 fpcr;
-	__uint128_t vregs[32];
+	__u64 vregs[32];
 };
 
 /* ESR_EL1 context */
